WTF babel you're drunk
The call in line 126 fails because args is an array with a function in
it. But if we access args later in the code (e.g. line 132, debug),
then it works correctly passing the true function arguments.

Unfortunately there’s also a bug in runtime that forces us to update to
a later version of Babel which contains this bug.